--- src/views/trip-info-container.js
@@ -1,15 +1,0 @@
-import Abstract from './abstract';
-
-const createTripInfoTemplate = () => '<section class="trip-main__trip-info  trip-info"></section>';
-
-export default class TripInfo extends Abstract{
-  constructor() {
-    super();
-  }
-
-  getTemplate () {
-    return createTripInfoTemplate();
-  }
-}
-
-
